requirements.txt

exa_py==1.7.0
fpdf2==2.8.1
google==3.0.0
googlesearch_python==1.2.5
groq==0.13.0
Markdown==3.7
markdown2==2.5.2
markdown_pdf==1.3.1
pdfkit==1.0.0

phidata==2.6.6
praw==7.8.1
protobuf==5.29.1
pydantic==2.10.3
pypdf==5.1.0
python-dotenv==1.0.1
Requests==2.32.3
rich==13.9.4
spacy==3.8.3

streamlit==1.41.0
tavily_python==0.5.0
together==1.3.5

pycountry
sqlalchemy